requirements_phase3.txt
# >=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>==
# Phase 3 Requirements
# 심리상담 AI - 국제화 및 고도화
# 저장 경로: /AI_Drive/counseling_ai/requirements_phase3.txt
# >=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>=>==

# -----------------------------------------------------------------------------
# Phase 1 & 2 의존성 포함
# -----------------------------------------------------------------------------
-r requirements.txt
-r requirements_phase2.txt

# -----------------------------------------------------------------------------
# 다국어 처리
# -----------------------------------------------------------------------------
# 언어 감지
langdetect>=1.0.9
fasttext>=0.9.2
lingua-language-detector>=2.0.2

# 다국어 NLP
konlpy>=0.6.0  # 한국어
mecab-python3>=1.0.8  # 일본어
jieba>=0.42.1  # 중국어
underthesea>=6.8.0  # 베트남어

# 번역
deep-translator>=1.11.4
googletrans>=4.0.0-rc1

# 다국어 임베딩
sentence-transformers>=2.2.2
# multilingual-e5-large>=0.1.0

# -----------------------------------------------------------------------------
# AI 슈퍼바이저
# -----------------------------------------------------------------------------
# 품질 평가
rouge-score>=0.1.2
bert-score>=0.3.13
sacrebleu>=2.3.1

# 안전성 분석
detoxify>=0.5.1
# perspective-api-client>=0.1.0

# 임상 평가 도구
scipy>=1.11.0
statsmodels>=0.14.0
pingouin>=0.5.3

# -----------------------------------------------------------------------------
# 개인화 및 추천
# -----------------------------------------------------------------------------
# 벡터 DB
pinecone-client>=2.2.4
milvus>=2.3.0
chromadb>=0.4.0

# 추천 시스템
implicit>=0.7.2
lightfm>=1.17

# 시계열 분석
prophet>=1.1.4
neuralprophet>=0.6.2

# -----------------------------------------------------------------------------
# 연구 플랫폼
# -----------------------------------------------------------------------------
# 통계 분석
lifelines>=0.27.7  # 생존 분석
pymc>=5.8.0  # 베이지안 분석
arviz>=0.16.1

# 데이터 익명화
anonymize-it>=0.0.3
sdv>=1.4.0  # 합성 데이터 생성

# 데이터 내보내기
pyreadstat>=1.2.2  # SPSS/Stata
openpyxl>=3.1.2
xlsxwriter>=3.1.2

# 시각화
plotly>=5.17.0
altair>=5.1.2
seaborn>=0.12.2

# -----------------------------------------------------------------------------
# 국제화
# -----------------------------------------------------------------------------
# 통화 및 지역화
babel>=2.12.1
py-moneyed>=3.0
python-money>=1.0.0

# 타임존
pytz>=2023.3
tzdata>=2023.3

# 국제화 프레임워크
django-modeltranslation>=0.18.11
flask-babel>=3.1.0

# -----------------------------------------------------------------------------
# 규정 준수 및 보안
# -----------------------------------------------------------------------------
# 암호화
cryptography>=41.0.0
pynacl>=1.5.0
python-jose[cryptography]>=3.3.0

# 감사 로깅
audit-log>=0.0.1
python-json-logger>=2.0.7

# GDPR/개인정보
consent-manager>=0.1.0
data-retention>=0.1.0

# -----------------------------------------------------------------------------
# 인프라 및 모니터링
# -----------------------------------------------------------------------------
# 분산 추적
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-exporter-jaeger>=1.20.0

# APM
elastic-apm>=6.18.0
datadog>=0.47.0
newrelic>=9.0.0

# 메트릭
prometheus-fastapi-instrumentator>=6.1.0
statsd>=4.0.1

# 헬스체크
py-healthcheck>=1.10.1

# -----------------------------------------------------------------------------
# 테스트 (Phase 3)
# -----------------------------------------------------------------------------
# 다국어 테스트
pytest-localserver>=0.8.0
responses>=0.23.3

# 부하 테스트
locust>=2.16.1
k6>=0.1.0

# 카오스 엔지니어링
chaostoolkit>=1.15.0
chaostoolkit-kubernetes>=0.29.0

# -----------------------------------------------------------------------------
# 개발 도구
# -----------------------------------------------------------------------------
# 문서화
mkdocs>=1.5.2
mkdocs-material>=9.2.0
mkdocstrings>=0.22.0

# API 문서
redoc>=2.0.0
swagger-ui-bundle>=0.0.9

# 코드 품질
ruff>=0.0.291
mypy>=1.5.1
bandit>=1.7.5
